fix: unify Int/UInt return types for getter consistency
- Wave field getters (frame_count, sample_rate, sample_size, channels):
  UInt → Int to match all other count/size getters (Image, Texture, Mesh, etc.)
- Shader::id(): Int → UInt to match Texture::id() and RenderTexture::id()
  (raylib's Shader.id is unsigned int)
- Update C stubs accordingly (int32_t for Wave, uint32_t for Shader)
- Remove now-redundant .reinterpret_as_uint() calls in shader examples

+++ internal/raylib/stub_audio.c
@@ -201,24 +201,24 @@ moonbit_raylib_load_wave_samples(WaveWrapper *w) {
   return r;
 }
 
-unsigned int
+int32_t
 moonbit_raylib_get_wave_frame_count(WaveWrapper *w) {
-  return w->data->frameCount;
+  return (int32_t)w->data->frameCount;
 }
 
-unsigned int
+int32_t
 moonbit_raylib_get_wave_sample_rate(WaveWrapper *w) {
-  return w->data->sampleRate;
+  return (int32_t)w->data->sampleRate;
 }
 
-unsigned int
+int32_t
 moonbit_raylib_get_wave_sample_size(WaveWrapper *w) {
-  return w->data->sampleSize;
+  return (int32_t)w->data->sampleSize;
 }
 
-unsigned int
+int32_t
 moonbit_raylib_get_wave_channels(WaveWrapper *w) {
-  return w->data->channels;
+  return (int32_t)w->data->channels;
 }
 
 // ============================================================================
